requirements.txt
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
smbus2==0.4.3
spidev==3.6
tzdata==2024.2
